Scatter-pie in Brazil Map
Fig. 1. The state-level proportion of total production amounts (top-right) and production yields (tons per hectare) of 45 leading crops throughout Brazil in 2017. Crops were categorized by pollinator dependence. Pie sizes represent production amounts (in tons) of the 45 crops; slice sizes represent average yields broken-down by the degree of pollinator dependence. The pie chart on top-right represents the proportion of pollinator dependence on the total production amount in 2017. Pie charts are color-coded according to the degree of pollinator dependence. Source: IBGE, 2020. (For interpretation of the references to color in this figure legend, the reader is referred to the Web version of this article.)